api-real/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
anthropic==0.8.0
playwright==1.41.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
python-telegram-bot==20.8
aiohttp==3.9.1
pydantic==2.6.0
schedule==1.2.0
redis==5.0.1
python-dotenv==1.0.0
